Factored out makeGutterPointer() into separate module, added smoke test.
